+++ upcloud/testdata/upcloud_kubernetes/kubernetes_s1.tf
@@ -1,10 +1,10 @@
 variable "name" {
-  default = "tf-acc-test"
+  default = "tf-acc-test-uks"
   type    = string
 }
 
 variable "zone" {
-  default = "de-fra1"
+  default = "fi-hel2"
   type    = string
 }
 
@@ -12,7 +12,7 @@ resource "upcloud_network" "main" {
   name = "terraform-provider-upcloud-test"
   zone = var.zone
   ip_network {
-    address = "172.23.5.0/24"
+    address = "172.23.7.0/24"
     dhcp    = true
     family  = "IPv4"
   }