+++ upcloud/testdata/upcloud_managed_database/redis_properties_s1.tf
@@ -1,5 +1,5 @@
 resource "upcloud_managed_database_redis" "redis_properties" {
-  name = "redis-properties-test"
+  name = "tf-redis-properties-test"
   plan = "1x1xCPU-2GB"
   zone = "fi-hel2"
   properties {
@@ -11,7 +11,7 @@ resource "upcloud_managed_database_redis" "redis_properties" {
     redis_pubsub_client_output_buffer_limit = 128
     redis_ssl                               = false
     redis_lfu_log_factor                    = 11
-    redis_io_threads                        = 2
+    redis_io_threads                        = 1
     redis_maxmemory_policy                  = "allkeys-lru"
     redis_persistence                       = "off"
     redis_timeout                           = 310