+++ upcloud/testdata/upcloud_server_group/step2.tf
@@ -14,9 +14,9 @@ resource "upcloud_server" "test" {
 }
 
 resource "upcloud_server_group" "tf_test_1" {
-  title         = "tf_test_1_updated"
-  anti_affinity = true
-  labels = {
+  title                = "tf_test_1_updated"
+  anti_affinity_policy = "strict"
+  labels               = {
     "key1" = "val1"
     "key2" = "val2"
   }